chore: add common module, move vector + ratio
