+++ go/common/go.mod
@@ -1,0 +1,3 @@
+module common
+
+go 1.17
