+++ go/common/ratio/ratio.go
