+++ go/common/ratio/ratio_test.go
