+++ go/common/sigfig.go
@@ -1,0 +1,3 @@
+// utilities for working with significant figures
+
+package common
