+++ go/common/vector/vector.go
